Add server and README.md (#5)
* moved  firewalls, added basic server class and added basic exec code validation.

* creating a udp listener

* formatting updated

* small changes

* move functions to the bottom

* update sequence to have ports as 16 bit number

* starting server

* adding  server and multi threaded listeners

* update firewall interface and set port to uint16_t

* remove warning

* add a central event to wake all threads

* removed warning

* remove extras

* Add Readme and remove invalid config

+++ src/sequence.h
@@ -7,8 +7,8 @@
 class Sequence {
 private:
   bool tcp; // false is udp
-  int unlockPort;
-  std::vector<int> knockPorts;
+  uint16_t unlockPort;
+  std::vector<uint16_t> knockPorts;
 
 public:
   // Check if the port is in range of actual ports and pushes back if valid
@@ -21,9 +21,9 @@ public:
   void setUnlockPort(const int port, const bool isTcp);
   // returns if sequence is valid
   bool isValid() const;
-  int getUnlockPort() const;
+  uint16_t getUnlockPort() const;
   bool isTcp() const;
-  std::vector<int> getKnockPorts() const;
+  std::vector<uint16_t> getKnockPorts() const;
   ~Sequence() = default;
 };
